Move works as an array tests to their own class
To reduce the size of teh ArrayMetaTest class, we move the tests for
checking that it can function as a normal array to its own class.

Anything that is its own thing, that can not be extracted from the
array meta class, should be in its own test class.
